changed KeySignature.fifthsAboveC from Int to Long

+++ src/main/kotlin/core/Score.kt
@@ -31,7 +31,7 @@ class Note(
 
 class KeySignature(
     anchorTick: Long,
-    val fifthsAboveC: Int,
+    val fifthsAboveC: Long,
 ) : StaffSymbol(anchorTick) {
 }
